Automatically resolve shlibs
* Use packages for build-deps and base runtime-deps
* shlibs are installed using traditional apt-get
